requirements.txt
requests>=2.31.0
mcp>=1.26.0
python-dotenv>=1.0.0
fastmcp>=3.2.0